Fix Label and IconLabel classes
When updating the sprites these labels didn't work so well. I was
deleting and recreating all of the sprites, which causes issues
(position gets lost, etc). Simplified the code, and now only the
material is updated when needed. Leaving it to the user to
manually ask for a .redraw() though.